adding filter view

+++ src/main/java/BGU/Group13B/service/ISession.java
@@ -195,7 +195,7 @@ public interface ISession {
      * @param minPrice - the minimum price of the product
      * @param maxPrice - the maximum price of the product
      */
-    Response<List<ProductInfo>> filterByPriceRange(int minPrice, int maxPrice);
+    Response<List<ProductInfo>> filterByPriceRange(double minPrice, double maxPrice);
 
     /**
      * #18
@@ -204,7 +204,7 @@ public interface ISession {
      * @param minRating - the minimum rating of the product
      * @param maxRating - the maximum rating of the product
      */
-    Response<List<ProductInfo>> filterByProductRank(int minRating, int maxRating);
+    Response<List<ProductInfo>> filterByProductRank(double minRating, double maxRating);
 
     /**
      * #18
@@ -221,7 +221,7 @@ public interface ISession {
      * @param minRating - the minimum rating of the store
      * @param maxRating - the maximum rating of the store
      */
-    Response<List<ProductInfo>> filterByStoreRank(int minRating, int maxRating);
+    Response<List<ProductInfo>> filterByStoreRank(double minRating, double maxRating);
 
     /**
      * #16
